Organise code in a src layout to allow relative imports work properly for running testflinger server application
