+++ server/src/testflinger/static/assets/js/filter.js
