+++ server/src/testflinger/templates/_agents_table.html
