+++ server/src/testflinger/templates/_footer.html
