+++ server/src/testflinger/templates/agent_detail.html
