+++ server/src/testflinger/templates/agents.html
