+++ server/src/testflinger/templates/job_detail.html
