+++ server/src/testflinger/templates/queue_detail.html
