+++ server/src/testflinger/templates/queues.html
