+++ server/tox.ini
@@ -26,6 +26,9 @@ commands =
 
 [testenv:lint]
 description = Run linting tests
+dependency_groups = 
+    dev
+    charm
 set_env =
     PYTHONPATH = {tox_root}/charm/src:{tox_root}/charm/lib
 commands =
